==> notebooks/1-Download-Data-And-Cleaning.ipynb <==
## Here's a notebook of how I cleaned the data I used in my example Bokeh app, VirtualDive
**NOTE: This notebook is optional: I've already provided cleaned data files in the data folder of this repo so that this does not need to be run before the Bokeh notebook**

1) Let's import our packages! For this bit all I used was Pandas and Numpy:

import pandas as pd
import numpy as np

2) Let's load our data into a couple Pandas dataframes! ReefLocations is actually from http://reefbase.org/ and the SEDAC 2000 population density data is actually from https://neo.sci.gsfc.nasa.gov/view.php?datasetId=SEDAC_POP . I've just reuploaded both to GitHub to make sure the files don't move or change in the future.

reeflocations=pd.read_csv("https://raw.githubusercontent.com/mistergroot/VirtualDive/master/data/ReefLocations.csv")
pop = pd.read_csv("https://raw.githubusercontent.com/mistergroot/VirtualDive/master/data/SEDAC_POP_2000-01-01_rgb_1440x720.SS.CSV")

3) Let's see what we have to work with:

reeflocations.head(4)

pop.head(4)

4) Let's start with the pop data. It looks pretty ugly. It's a gridded dataset where the top row and left column are longitudes and latitudes, respectively. I don't like that the latitudes are in an unindexed column. Let's change that. It's also going to make the index look weird with the name 'lat/lon' there, so let's get rid of it at the same time:

pop = pop.set_index('lat/lon')
del pop.index.name

pop.head(2)

5) Okay now let's go ahead and convert this gridded data to "lat", "lon", and "popdens" columns. This line of code essentially finds all `'item'`s in the dataframe and displays its column and row name side by side. I wanted data greater than 250 people per km^2 (because if we start at 0 we'll have maybe 100,000 datapoints and that's excessive for my needs) and less than 99999 (because 99999 is a filler value which is equivalent to no data/no people):

data = pop.stack().to_frame('item').query('250 < item < 99999')

data.head(4)

6) Looks good! Formatting is a bit messed up though, so let's reset the indexed column and rename the columns to the correct thing.

data = data.reset_index(drop=False)
data.rename(columns={'level_0':'lat','level_1':'lon','item':'popdens'}, inplace=True)
data.head(4)

7) We're done with the population data! Let's export it to a CSV:

data.to_csv("../data/popdata.csv")

8) Now let's start on the reef data. This should be easier to do since it's already in columns rather than gridded. Firstly let's check what datatype our columns are in:

reeflocations.dtypes

9) I don't like that some numeric data is stored as objects (e.g. WATER_DEPTH) so we'll convert objects that are numeric into floats. We're not actually going to use the water depth later, I just want to do this anyway:

reeflocations = reeflocations.convert_objects(convert_numeric=True)
reeflocations.dtypes

10) Now, in the unformatted data, unprotected reefs are represented with 0 and protected reefs are represented with a 1. If data is absent, I'd like to assume that they're not protected, so here we'll use `fillna` to add 0s to the missing portions.

reeflocations['PROTECTED'] = reeflocations['PROTECTED'].fillna(0)

11) I want to replace the 1s and 0s with "Yes" and "No" now simply because that's how I want it displayed in the glyphs later. I had trouble doing this with the Boolean values of 1 and 0 for some reason so I just convert it to an int first.

reeflocations['PROTECTED'] = reeflocations['PROTECTED'].astype(np.int64)
reeflocations["PROTECTED"] = reeflocations["PROTECTED"].replace(1, "Yes")
reeflocations["PROTECTED"] = reeflocations["PROTECTED"].replace(0, "No")
reeflocations.head(4)

12) Looks good! Let's split up the data into two separate dataframes based on whether or not they're protected:

protected = reeflocations[reeflocations["PROTECTED"] == "Yes"]
unprotected = reeflocations[reeflocations["PROTECTED"] == "No"]

13) And finally, we'll export the split and unsplit reef data to CSVs using Pandas:

reeflocations.to_csv("../data/reefloc.csv")
protected.to_csv("../data/protected.csv")
unprotected.to_csv("../data/unprotected.csv")

## And we're done processing the raw data! Now to move on to the actual Bokeh app
