==> notebooks/2-Example-Bokeh-App-VirtualDive.ipynb <==
## Here we import our required packages
These are packages which are used later in the formatting of our interactive Bokeh plots but we need to import them now before we use them. These will be introduced in the coming cells.

## These are a lot of the tools which we'll use to build our plots
from bokeh.models import (
  GMapPlot, GMapOptions, ColumnDataSource, Circle, Annulus, Legend, LegendItem,
    Range1d, PanTool, WheelZoomTool, HoverTool, TapTool, CustomJS
)
## These are the colormappers we'll use to colorcode our glyphs (points on the interactive map)
from bokeh.models.mappers import CategoricalColorMapper, LogColorMapper

## This is the palette I chose to use but you can use any color palettes that fit your needs.
## Bokeh has many color palettes included but I prefered those of colorcet
from colorcet import bkr as palette, fire as fire

## These tools show and output our Bokeh plot as an HTML file
from bokeh.io import output_file, output_notebook, show

## Pandas is used to import our formatted data CSV files into dataframes usable by ColumnDataSource
import pandas as pd

## Reading CSV data as Pandas dataframes
Here were use pandas to load our data that we downloaded and cleaned in `DataCleaning.ipynb`. `Protected` is protected reefs and their associated data, `unprotected` is unprotected reefs and their associated data, and `data` is NASA SEDAC population density data sorted by >250 people per km^2 with a spatial resolution of 0.25 degrees (from the year 2000. Future iterations of VirtualDive will include current populations as well as projections).

protected = pd.read_csv("../data/protected.csv")
unprotected = pd.read_csv("../data/unprotected.csv")
data = pd.read_csv("../data/popdata.csv")

This is not necessary but I always like to view what my data looks like when it's being read, just in case I added the wrong file along the way:

protected.head(4)

unprotected.head(4)

data.head(4)

## Now we start the fun bit
Here we start setting up our preferences for the Google Maps API using the GMapOptions we imported earlier

map_options = GMapOptions(lat=7.2, ## Here I choose the latitude and longitude that I want the map
                          lng=100, ## to be centered. I chose the Coral Triangle as the center.
                          scale_control=True, ## Do we want Google Maps API to display a scale bar?
                          map_type="satellite", ## Options: "roadmap", "satellite", "terrain", "hybrid"
                          zoom=4) ## Set the zoom level to what is relevant for your needs

## Defining our data that will be used by Bokeh
Here I define 3 ColumnDataSources which utilize the CSV files we read as dataframes earlier and will be used by Bokeh when creating glyphs in the next few cells.

protectedsource = ColumnDataSource(
    data=dict(
        lat=protected['LAT'], ## Latitudes of the protected reefs
        lon=protected['LON'], ## Longitudes of the protected reefs
        color=protected['PROTECTED'], ## Glyphs will be colorcoded later by protected status
                                        ## so we name the status as "color"
        reefnames=protected.REEF_NAME.tolist(), ## Names of reefs are stored as lists
        reeftype=protected.REEF_TYPE.tolist(), ## Reef types are stored as lists
    )
)

unprotectedsource = ColumnDataSource(
    data=dict(
        lat=unprotected['LAT'], ## Latitudes of the unprotected reefs
        lon=unprotected['LON'], ## Longitudes of the unprotected reefs
        color=unprotected['PROTECTED'], ## Glyphs will be colorcoded later by protected status
                                        ## so we name the status as "color"
        reefnames=unprotected.REEF_NAME.tolist(), ## Names of reefs are stored as lists
        reeftype=unprotected.REEF_TYPE.tolist(), ## Reef types are stored as lists
    )
)

popsource = ColumnDataSource(
    data=dict(
        poplat=data['lat'], ## Latitudes of each population density datapoint
        poplon=data['lon'], ## Longitudes of each population density datapoint
        colorpop=data['popdens'], ## Population densities found at each coordinate
    )
)

## Defining our colormappers
Here we define two colormappers, one categorical and one logarithmic. A categorical colormapper was used because reefs will be colored by whether they are protected or not. In the case of my map, the factor 'No' corresponds with a magenta color and the factor 'Yes' corresponds with a cyan color. The colors can be changed to whatever colors you'd like. The population density colormapper is logarithmic because the population densities follow an exponential distribution. The LogColorMapper linearizes the palette.

reefcolormapper = CategoricalColorMapper(palette=["magenta", "cyan"], factors=['No','Yes'])
popcolormapper = LogColorMapper(palette=fire) ## This is the colorcet palette I 
                                                ## decided to use instead of the Bokeh palettes

## Defining our glyphs (interactive datapoints)
Here we set up the way in which our individual glyphs will be displayed on the map. I used annuli (rings) for the reef locations and circles for the population data points. I think this section is fairly straightforward but I'll explain it more within the cells.

protectedannulus = Annulus(x="lon", ## The annulus is centered at this X coordinate (longitude)
                y="lat", ## The annulus is centered at this Y coordinate (latitude)
                inner_radius=4000, ## Inside radius size of the annulus
                outer_radius=7000, ## Outside radius size of the annulus
                fill_color={'field': 'color', 'transform': reefcolormapper}, ## Here we specify
                           ## which of the previously defined colormappers we're using for this
                           ## set of glyphs. Here we used reefcolormapper.
                fill_alpha=0.3, ## Since we have many datapoints overlapping I wanted a very
                           ## light opacity so I set it to an alpha of 0.3.
                line_color=None) ## I didn't want any outlines around the annulus so I set this
                                ## to none. This is all up to personal preference.

unprotectedannulus = Annulus(x="lon", 
                y="lat", 
                inner_radius=4000,
                outer_radius=7000,
                fill_color={'field': 'color', 'transform': reefcolormapper},
                fill_alpha=0.3, 
                line_color=None)

popcircle = Circle(x="poplon", 
                y="poplat", 
                radius=12000, ## Note: size can be used as a replacement for radius. Glyphs scale
                               ## with zoom level if using radius but do not scale when using
                               ## size. Glyphs would remain the same size regardless of zoom level.
                fill_color={'field': 'colorpop', 'transform': popcolormapper}, 
                fill_alpha=0.8,
                line_color=None)

## Setting up tools for user interactions
In this example app I've chosen to use the TapTool, the WheelZoomTool, the PanTool, and the HoverTool. If you'd like to learn more about these tools or about what other tools you can utilize, follow this link:

https://bokeh.pydata.org/en/latest/docs/user_guide/tools.html

## Here we're going to define a JavaScript callback using Bokeh's CustomJS feature.
## As of right now, all this does is make it so when you click a glyph using the
## tap tool, it reduces the opacity of the non-selected glyphs. In the future this
## will be developed further to show the species pool of each selected reef in a table
## below the map. Implementing a live-updating table below the map ended up being
## too complicated to meet the project deadline but will be included in the final
## version of VirtualDive.
cb_code = """
var ind = source.selected['1d'].indices;
var data = source.data;
console.log(source['z'][ind])
"""

tap = TapTool(name="foo", ## "foo" refers to the glyph that the taptool is interacting
                          ## with. You can set this to whatever you want as long as it
                          ## corresponds with the same name found in glyphs. You'll see
                          ## the "foo" name reappear when we actually add the glyphs in
                          ## the next few cells.
              callback = CustomJS(args=dict(source=protectedsource), ## Here we set the
                                                    ## source of the JavaScript callback
                                  code = cb_code)) ## and add the callback code.

zoom = WheelZoomTool() ## We don't really need to change anything here since all this
                        ## tool does is zoom in and out using the scroll wheel of a
                        ## mouse or two finger scroll of a touchpad.

pan = PanTool(dimensions="both") ## Here we can specify which directions we want the pan
                                ## tool to work in. I want the user to be able to pan
                                ## side to side and up and down so I set this as "both"


## Here we set the hover tool to show only on glyphs with the name "foo". We also set
## how we want the popups to appear using HTML. I have it set here so that the box is
## black and the text is white for name and type and #696 for GPS coordinates. We make
## sure the hovertool shows the right data by setting the hovertool to make use of 
## @reefnames, @reeftype, and @lat, @lon from the ColumnDataSources of the glyphs.
hover = HoverTool(names=["foo"], tooltips="""
    <HTML>
    <HEAD>
    <style>    
    .bk-tooltip {
        background-color: black !important;
        }
    </style>
    </HEAD>
    <BODY>
    <div class = ".bk-tooltip">
        <div>
            <span style="font-size: 12px; font-weight: bold; color: white;">@reefnames</span>
        </div>
        <div>
            <span style="font-size: 10px;color: white;">@reeftype Reef</span>
        </div>
        <div>
            <span style="font-size: 10px; color: #696;">(@lat, @lon)</span>
        </div>
    </div>
    </BODY>
    </HTML>
    """)

## Setting up the plot and actually implementing glyphs
Here we go. Actually generating the plot. Finally.

plot = GMapPlot(
    x_range=Range1d(-160, 160), ## Google Maps API is sort of weird. I've set these
    y_range=Range1d(-80, 80), ## bounds with Range1d from Bokeh but they don't seem to
                                ## work how I want but the map doesn't work without it
    
    map_options=map_options, ## Here we use the map options we defined earlier
    
    sizing_mode='stretch_both', ## This means that the plot will stretch to fill whatever
                                ## space is available to it. When you embed the plot later
                                ## it will fill whatever space on your website you give it.
    
    tools=[hover, tap, zoom, pan] ## Here we just say which tools we want to use (all the
                                    ## tools we defined in the previous step)
)

plot.title.text = "VirtualDive" ## What title do we want?
plot.title.text_font_size = "25px" ## How big do we want it?
plot.title_location="right" ## Where do we want it?
plot.title.align = "right" ## How do we want it to align wherever it is?

## To run GMapPlot, you need to include a Google Maps API key. If you make a map of your
## own, please get a key from:
## https://developers.google.com/maps/documentation/javascript/get-api-key
## I've included mine here for the tutorial but please don't use it to build
## your own Bokeh app.
plot.api_key = "AIzaSyAX0RhQ5JTdQAjveEADHzBXbxkVLYCiPps"

## Let's add the glyphs we defined earlier. Each glyph has to have the data source and
## glyph. Optionally you can include a name so that tools will only be active for the names
## that they match (for this example the names are "foo").
plot.add_glyph(protectedsource, protectedannulus, name="foo")
plot.add_glyph(unprotectedsource, unprotectedannulus, name="foo")
plot.add_glyph(popsource, popcircle)

## Adding a glyph legend
You thought we were done, didn't you? Well technically we are. You could run `show(plot)` and it would spit out a completed Bokeh app, but we want to add a legend. For this map we'll just add a legend for the reef glyphs. I'll be adding a colorbar with tickmarks for the population glyphs but for now I'm leaving it out because I'm still figuring that out myself.

## Here we set up our LegendItems. We set a name with "label" and say which renderer it
## corresponds with. Protected Reef corresponds with renderer 0 since we defined that glyph
## first in the previous step. Unprotected Reef corresponds with renderer 1.
protectedreef = LegendItem(label='Protected Reef', renderers=[plot.renderers[0]])
unprotectedreef = LegendItem(label='Unprotected Reef', renderers=[plot.renderers[1]])

## Now that we have the individual LegendItems set up, we lump them into a Legend and refer
## back to them using "items". I want the legend at the top right so I set the location
## to "top_right".
legend = Legend(items=[protectedreef, unprotectedreef], location='top_right')

## And finally we add the legend to our plot using add_layout
plot.add_layout(legend)

## Viewing the fruits of your labor!
We're done! Run `show(plot)` and it'll open your Bokeh app in a new tab. To save it, check out the next cell. <span style="color:red">**NOTE: Because of an issue with the Google Maps API, the 180th Meridian (around Fiji) cannot be in the frame otherwise the glyphs all shift or disappear. If you pan towards Hawaii, the glyphs will disappear until the 180th Meridian is out of frame. Please keep this in mind when interacting with this example Bokeh app. This is a limitation with Google Maps API, not with Bokeh. I will fix this by switching over to a WMTS XYZ Tile Source in the future.**</span>

show(plot)

If you'd like to save and view your Bokeh map as an HTML file, you can include these lines of code. I didn't run this so there's no saved HTML file in the GitHub repo. Run it if you want to view VirtualDive as an HTML.

output_notebook()
output_file("VirtualDive.html")

## If you'd like to see how to use Heroku to host a Bokeh app server, see the next tutorial.
