==> notebooks/3-Embedding-With-Heroku.ipynb <==
## Let's try embedding my Bokeh app to a website using a Heroku server
NOTE: This notebook doesn't actually have any runnable code in it. I just wanted all my tutorials to be in Jupyter Notebooks.

## Firstly, what is Heroku?
From Wikipedia: Heroku is a cloud platform as a service (PaaS) supporting several programming languages that is used as a web application deployment model. Heroku, one of the first cloud platforms, has been in development since June 2007, when it supported only the Ruby programming language, but now supports Java, Node.js, Scala, Clojure, Python, PHP, and Go.

## Step 1: Install Heroku's command-line interface on your system
Follow the steps in this link on Heroku's website:
https://devcenter.heroku.com/articles/heroku-cli

## Step 2: Signup for a free account with Heroku
Create an account at: https://www.heroku.com/

## Step 3: Setting up things on your end
1) Login to Heroku in your terminal by running

`heroku login`

2) Create a new Heroku app by running 

`heroku create appname` where `appname` is the name of the app you'd like to create

3) In your file manager, create an empty folder wherever you'd like and name it whatever you called your app. Now navigate to it in your file manager. Also navigate to it in your terminal using:

`cd PATH` where `PATH` is the path to your new app folder

## Step 4: Collecting and moving files
If you've cloned my entire repo and you're running this tutorial from within it, move the following files to the folder you created in the previous step:

1) `../data/popdata.csv`

2) `../data/protected.csv`

3) `../data/unprotected.csv`

4) `../project/virtualdive.py`

5) `../project/requirements.txt`

6) `../project/runtime.txt`

7) `../project/Procfile`

If you're just following this tutorial online and haven't cloned my repo, just download the above files from my GitHub and put them in the folder you created earlier.

## Step 5: Let me explain what's actually in those files so that you can create your own embedded app

1) `../data/popdata.csv`

This is simply a data file that is read by `virtualdive.py`. We created this in the first tutorial notebook.

2) `../data/protected.csv`

This is simply a data file that is read by `virtualdive.py`. We created this in the first tutorial notebook.

3) `../data/unprotected.csv`

This is simply a data file that is read by `virtualdive.py`. We created this in the first tutorial notebook.

4) `../project/virtualdive.py`

This is a modified version of the example Bokeh app we created in the first tutorial notebook. What's different? Great question. In our imports, instead of importing:

## These tools show and output our Bokeh plot as an HTML file
from bokeh.io import output_file, output_notebook, show

We import:

from bokeh.io import curdoc

This is used by the server to display your app. `output_file`, `output_notebook`, and `show` aren't necessary because we're not saving an HTML file when we run it on the server. The import section of the .py file looks like this:

from bokeh.io import curdoc
from bokeh.models import (
  GMapPlot, GMapOptions, ColumnDataSource, Circle, Annulus, Legend, LegendItem,
    Range1d, PanTool, WheelZoomTool, CustomJS, HoverTool, TapTool
)
from bokeh.models.mappers import CategoricalColorMapper, LogColorMapper
from colorcet import bkr as palette, fire as fire
import pandas as pd
import numpy as np`

Additionally, compared to the Bokeh script in the Jupyter Notebook example, loading our data is a bit different. If you'd like, you can make the .py file reference to subfolder in Heroku, but I like having everything in a single directory when there's only a few files like we have here. If you have a more complex app, you'd want to add subfolders similar to how this repository is organized. Here's how we're adding data:

protected = pd.read_csv("protected.csv")
unprotected = pd.read_csv("unprotected.csv")
data = pd.read_csv("popdata.csv")

Lastly, at the end of the function, we no longer have `show` or `out_file`. Instead we end the .py with:

curdoc().add_root(plot)

This tells the server to display our plot

5) `../project/requirements.txt`

The `requirements.txt` file simply tells Heroku which packages are required to run your script. In my case, file looks like this on the inside:

`bokeh==0.12.15
numpy==1.14.2
pandas==0.22.0
colorcet==0.9.1`

It's good to specify the version that you've tested on your system so that you know it will work on Heroku's servers.

6) `../project/runtime.txt`

The `runtime.txt` file tells Heroku which Python environment to run your script in. The inside of the file looks like:

`python-3.6.5`

In my example, we need Python 3.6.5 but you can add whatever version of Python you've built your script in as long as it works on Heroku's servers.

7) `../project/Procfile`

Finally, the `Procfile` is a file with **NO FILE EXTENSION** and the first letter **MUST BE CAPITALIZED**. I don't know much about what the `Procfile` actually means but here's what it looks like inside for my example:

`web: bokeh serve --log-level=debug --port=$PORT --num-procs=0 --host=virtualdive.herokuapp.com --address=0.0.0.0 --use-xheaders virtualdive.py`

In order to run this with your app, you need to change `virtualdive.herokuapp.com` to `appname.herokuapp.com` where `appname` is the name of your app you created in **Step 3**. You also need to change `virtualdive.py` to whatever you named your script. If you're just learning based on my example before creating your own script and app, leave it as `virtualdive.py`.

## Step 5: Sending your files over to Heroku
1) In terminal, you should have already navigated to your folder with the above files in it. If not, run

`cd PATH`

Where `PATH` is the path to the folder containing the files described in **Step 4** and **5**.

2) In terminal, run:

`git init`

This will initialize the folder with a `.git` file.

3) In terminal, run:

`git add .`

This will `git add` all the files within the current directory.

4) In terminal, run:

`git commit -m "Initial"`

This will commit all the files that you just `git add`ed. The `"Initial"` part can be whatever you want. It's simply a descriptor of the commit.

5) In terminal, run:

`git push heroku master`

This will push all your files in the current folder to your Heroku app.

6) In terminal, run:

`heroku open`

This will open your completed app in a new tab. It may take a minute or two to load the first time.

You're done! I'm going to provide an additional step of me embedding it into a embed block on my SquareSpace website but I'm not going to go into much detail just because I don't understand how it works.

## Step 6: Embedding in a SquareSpace code block

In order to embed the Bokeh app hosted on Heroku onto my website, I added an embed block to an empty webpage and added the following script by clicking the `</>` icon to the right:

     <iframe frameborder="0" style="border-width:0;" id="iframe" width="100%" height="920"></iframe>
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.3/jquery.min.js"></script>
     <script>
     (function($){
         $('#iframe').attr('src', 'https://virtualdive.herokuapp.com/virtualdive');
     })(jQuery);
     </script>
     
Again, I don't fully understand the specifics of what's going on in this block of code, but in order for it to work on your website you need to change `https://virtualdive.herokuapp.com/virtualdive` to the link of your app. Additionally, you can change the height and width of embedded block by altering the `width` and `height` portions of the code. I changed width to `100%` just so it fills whatever space I give it. The values here can be given as a percent or a pixel value (such as `920` which I gave as the `height` in my script).

## Enjoy the fruits of your labor!

My example is live at http://www.locatelliphotography.com/new-page-1

My Heroku account is free so I have a limited number of allowed awake hours per month so the server falls asleep from 12am-6am PST. If you go to the link above during that time, it may take ~10-20 seconds to load as the server wakes up from sleep.
